requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
mlflow==2.8.1
dvc==3.30.0
kagglehub==0.2.5
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
pyyaml==6.0.1
